refactor: adjust entities id types

+++ server_springboot/src/main/java/com/barataribeiro/taskr/repositories/entities/OrganizationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
+public interface OrganizationRepository extends JpaRepository<Organization, Long> {
     @Query("select distinct o from Organization o order by o.createdAt")
     Page<Organization> getAllOrganizationsPageable(Pageable pageable);
 }
