+++ server_springboot/src/main/java/com/barataribeiro/taskr/repositories/relations/OrganizationProjectRepository.java
@@ -12,9 +12,9 @@ import java.util.Set;
 public interface OrganizationProjectRepository extends JpaRepository<OrganizationProject, Long> {
     @EntityGraph(attributePaths = {"project"})
     @Query("select o from OrganizationProject o where o.organization.id = :id order by o.project.createdAt")
-    Set<OrganizationProject> findAllByOrganization_Id(Integer id);
+    Set<OrganizationProject> findAllByOrganization_Id(Long id);
 
-    Optional<OrganizationProject> findByOrganization_IdAndProject_Id(Integer id, Integer projectId);
+    Optional<OrganizationProject> findByOrganization_IdAndProject_Id(Long id, Long projectId);
 
     void deleteByProject(Project project);
